+++ app/javascript/vue/tasks/collection_objects/filter/utils/listParser.js
@@ -3,6 +3,12 @@ import { getDataAttributesFor } from '@/shared/Filter/utils'
 import { DataAttribute } from '@/routes/endpoints'
 import { flattenObject } from '@/helpers'
 
+function getTaxonDetermination(determinations) {
+  return determinations.length
+    ? determinations.toSorted((a, b) => a.position - b.position)[0]
+    : []
+}
+
 export async function listParser(list, { parameters }) {
   const { extend, exclude, ...rest } = parameters
   const { body } = await DataAttribute.brief({
@@ -38,9 +44,7 @@ export async function listParser(list, { parameters }) {
       current_repository,
       repository,
       collecting_event,
-      taxon_determinations: taxon_determinations.map((item) =>
-        flattenObject(item)
-      ),
+      taxon_determinations: getTaxonDetermination(taxon_determinations),
       dwc_occurrence,
       identifiers,
       data_attributes: getDataAttributesFor(body, item.id)