+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 ### Changed
 
 - Local identifier display in CollectionObject tag now prefers position to break tie with RecordNumber and CatalogNumber [#4074]
+- Filter CO: Show only current taxon determination [#4092]
 
 ### Fixed
 
@@ -26,12 +27,13 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Containerizing objects prevented identific increments
 - New dichotomous key: Radial annotator is not loading the correct data [#4076]
 
-[#4083]: https://github.com/SpeciesFileGroup/taxonworks/issues/4083
+[#970]: https://github.com/SpeciesFileGroup/taxonworks/issues/970
 [#4026]: https://github.com/SpeciesFileGroup/taxonworks/issues/4026
 [#4074]: https://github.com/SpeciesFileGroup/taxonworks/issues/4074
 [#4076]: https://github.com/SpeciesFileGroup/taxonworks/issues/4076
 [#4077]: https://github.com/SpeciesFileGroup/taxonworks/issues/4077
-[#970]: https://github.com/SpeciesFileGroup/taxonworks/issues/970
+[#4083]: https://github.com/SpeciesFileGroup/taxonworks/issues/4083
+[#4092]: https://github.com/SpeciesFileGroup/taxonworks/issues/4092
 
 ## [0.44.3] - 2024-10-03